refactor: Move data utils to 📁`Data`
Trying to keep all the data-related code in one place.
